+++ benchmarks/versus-external-pools/dynamic-worker-nodes.js
@@ -1,9 +1,6 @@
 // IMPORT LIBRARIES
-const { job, start } = require('microjob')
+const WorkerNodes = require('worker-nodes')
 // FINISH IMPORT LIBRARIES
-// IMPORT FUNCTION TO BENCH
-const functionToBench = require('./functions/function-to-bench')
-// FINISH IMPORT FUNCTION TO BENCH
 const size = Number(process.env.POOL_SIZE)
 const iterations = Number(process.env.NUM_ITERATIONS)
 const data = {
@@ -12,18 +9,19 @@ const data = {
   taskSize: process.env.TASK_SIZE
 }
 
+const workerNodes = new WorkerNodes(
+  require.resolve('./workers/worker-nodes/function-to-bench-worker'),
+  {
+    minWorkers: size,
+    maxWorkers: size * 3,
+    taskTimeout: 60000 // this is the same as poolifier default
+  }
+)
+
 async function run () {
-  await start({ maxWorkers: size })
   const promises = []
   for (let i = 0; i < iterations; i++) {
-    promises.push(
-      job(
-        data => {
-          functionToBench(data)
-        },
-        { data, ctx: { functionToBench } }
-      )
-    )
+    promises.push(workerNodes.call.functionToBench(data))
   }
   await Promise.all(promises)
   // eslint-disable-next-line no-process-exit