+++ benchmarks/versus-external-pools/fixed-poolifier.js
@@ -5,7 +5,8 @@ const size = Number(process.env.POOL_SIZE)
 const iterations = Number(process.env.NUM_ITERATIONS)
 const data = {
   test: 'MYBENCH',
-  taskType: process.env['TASK_TYPE']
+  taskType: process.env.TASK_TYPE,
+  taskSize: process.env.TASK_SIZE
 }
 
 const fixedPool = new FixedThreadPool(
@@ -19,6 +20,7 @@ async function run () {
     promises.push(fixedPool.execute(data))
   }
   await Promise.all(promises)
+  // eslint-disable-next-line no-process-exit
   process.exit()
 }