+++ benchmarks/versus-external-pools/pool-threadwork.js
@@ -6,4 +6,4 @@ const functionToBench = require('./functions/function-to-bench')
 // FINISH IMPORT FUNCTION TO BENCH
 const size = Number(process.env.POOL_SIZE)
 
-module.exports = new ThreadPool({ task: functionToBench, size: size })
+module.exports = new ThreadPool({ task: functionToBench, size })